Guild request async v2
